Added Grid and Square components
These components allough the board to be set. At this point the grid 
square hold one item of state whcih is undefined. In the next step I 
will create the objects containg the status for each square including 
whether they are emtpy, hold a ship or have taken a strike.

The square are child componets linked to the redux store so they will be 
re rendered if there state changes which makes things nice and DRY.