Ship Placement Page Added
This commited includes some refactoring of the way in which grids are 
allocated to the players. Previously players were initialised with two 
grids; a strikes grid and a ships grid but this has been changed so that 
each player has one grid with their ships. The strikes can be added to 
the opponents grid which can be whilst the ships remain hidden. To 
implement this change there have been some necessary amendments made to 
the GameActions and GameReducer as well as amending the object structure 
of the ships.

An unSelectedShips selector has been created enabling only ships that 
havent been placed to be available for placement.

The Square component now renders with state passed down from teh Grid. 
This method can be repeated in the next step of development when strikes 
will be introduced.

Some CSS styling has been added to the Square and ShipLlistItem 
components to help with ship placement. The syling classes are applied 
conditionally by the parent class.