Strikes can now be made
The Grid logic has been refactored into a Grid class which each player 
is instantiated with. Depending on whether the grids belongs to the 
player or to their opponent dictates what information the grid renders.

The GameReducer has a new PLACE_STRIKE action which assigsn the strike 
to the current players strikes.

The Square css has some new classes that render the square differently 
depending on whether it holds a ship, a strike, a ship that has been 
striked or nothing.

The GamePage renders with two grids. One for the player to see their 
ships and the strikes that have ben made against them. The second is for 
them to view the strikes they have made against their component 
including the succssful strikes but not the ships. The two grids will be 
instance of the same Grid class and therefore in order to identify them 
a 'type' property is now being passed to the class. e.g. 'owner' or 
'opponent'

Once a strike has been made the there is a 500 second delay for the 
player to see whether the strike was sucessful or not and the current 
player switches to he next player. To keep UX consistant the direction 
state is reset back to horizontal at this point.