Strikes can now be made
The Grid logic has been refactored into a Grid class which each player 
is instantiated with. Depending on whether a grid belongs to the player 
or to their opponent dictates what information the grid makes available.

The GameReducer has a new PLACE_STRIKE action which assigns the strike 
to the current players strikes.

The Square css has some new classes that render the square differently 
depending on whether it holds a ship, a strike, a ship that has received 
a strike or nothing.

The GamePage renders with two grids. One for the player to see their 
ships and the strikes that have ben made against them. The second is for 
them to view the strikes they have made against their opponent including 
the successful strikes but not the ships. The two grids will be instance 
of the same Grid class and therefore in order to identify them a 'type' 
property is now being passed to the class. e.g. 'owner' or 'opponent'

Once a strike has been made the there is a 500 millisecond delay for the 
player to see whether the strike was successful or not and the current 
player switches to he next player. To keep UX consistent the direction 
state is reset back to horizontal at this point.